Create common biometrics and pin unlock UI elements
