+++ app/src/test/java/com/x8bit/bitwarden/ui/platform/feature/settings/accountsecurity/AccountSecurityScreenTest.kt
@@ -23,6 +23,7 @@ import com.x8bit.bitwarden.data.platform.repository.model.VaultTimeoutAction
 import com.x8bit.bitwarden.data.platform.repository.util.bufferedMutableSharedFlow
 import com.x8bit.bitwarden.ui.platform.base.BaseComposeTest
 import com.x8bit.bitwarden.ui.platform.base.util.asText
+import com.x8bit.bitwarden.ui.platform.components.toggle.UnlockWithPinState
 import com.x8bit.bitwarden.ui.platform.manager.biometrics.BiometricsManager
 import com.x8bit.bitwarden.ui.platform.manager.intent.IntentManager
 import com.x8bit.bitwarden.ui.util.assertNoDialogExists
@@ -244,7 +245,11 @@ class AccountSecurityScreenTest : BaseComposeTest() {
             .performScrollTo()
             .performClick()
 
-        verify { viewModel.trySendAction(AccountSecurityAction.UnlockWithPinToggle.Disabled) }
+        verify {
+            viewModel.trySendAction(
+                AccountSecurityAction.UnlockWithPinToggle(UnlockWithPinState.Disabled),
+            )
+        }
     }
 
     @Suppress("MaxLineLength")
@@ -285,7 +290,11 @@ class AccountSecurityScreenTest : BaseComposeTest() {
             .filterToOne(hasAnyAncestor(isDialog()))
             .assertIsDisplayed()
 
-        verify { viewModel.trySendAction(AccountSecurityAction.UnlockWithPinToggle.PendingEnabled) }
+        verify {
+            viewModel.trySendAction(
+                AccountSecurityAction.UnlockWithPinToggle(UnlockWithPinState.PendingEnabled),
+            )
+        }
     }
 
     @Suppress("MaxLineLength")
@@ -304,7 +313,11 @@ class AccountSecurityScreenTest : BaseComposeTest() {
             .filterToOne(hasAnyAncestor(isDialog()))
             .performClick()
 
-        verify { viewModel.trySendAction(AccountSecurityAction.UnlockWithPinToggle.Disabled) }
+        verify {
+            viewModel.trySendAction(
+                AccountSecurityAction.UnlockWithPinToggle(UnlockWithPinState.Disabled),
+            )
+        }
         composeTestRule.assertNoDialogExists()
     }
 
@@ -324,7 +337,11 @@ class AccountSecurityScreenTest : BaseComposeTest() {
             .filterToOne(hasAnyAncestor(isDialog()))
             .performClick()
 
-        verify { viewModel.trySendAction(AccountSecurityAction.UnlockWithPinToggle.Disabled) }
+        verify {
+            viewModel.trySendAction(
+                AccountSecurityAction.UnlockWithPinToggle(UnlockWithPinState.Disabled),
+            )
+        }
         composeTestRule.assertNoDialogExists()
     }
 
@@ -368,7 +385,11 @@ class AccountSecurityScreenTest : BaseComposeTest() {
             .filterToOne(hasAnyAncestor(isDialog()))
             .assertIsDisplayed()
 
-        verify { viewModel.trySendAction(AccountSecurityAction.UnlockWithPinToggle.PendingEnabled) }
+        verify {
+            viewModel.trySendAction(
+                AccountSecurityAction.UnlockWithPinToggle(UnlockWithPinState.PendingEnabled),
+            )
+        }
     }
 
     @Suppress("MaxLineLength")
@@ -398,9 +419,11 @@ class AccountSecurityScreenTest : BaseComposeTest() {
 
         verify {
             viewModel.trySendAction(
-                AccountSecurityAction.UnlockWithPinToggle.Enabled(
-                    pin = "1234",
-                    shouldRequireMasterPasswordOnRestart = false,
+                AccountSecurityAction.UnlockWithPinToggle(
+                    UnlockWithPinState.Enabled(
+                        pin = "1234",
+                        shouldRequireMasterPasswordOnRestart = false,
+                    ),
                 ),
             )
         }
@@ -432,9 +455,11 @@ class AccountSecurityScreenTest : BaseComposeTest() {
 
         verify {
             viewModel.trySendAction(
-                AccountSecurityAction.UnlockWithPinToggle.Enabled(
-                    pin = "1234",
-                    shouldRequireMasterPasswordOnRestart = false,
+                AccountSecurityAction.UnlockWithPinToggle(
+                    UnlockWithPinState.Enabled(
+                        pin = "1234",
+                        shouldRequireMasterPasswordOnRestart = false,
+                    ),
                 ),
             )
         }
@@ -467,9 +492,11 @@ class AccountSecurityScreenTest : BaseComposeTest() {
 
         verify {
             viewModel.trySendAction(
-                AccountSecurityAction.UnlockWithPinToggle.Enabled(
-                    pin = "1234",
-                    shouldRequireMasterPasswordOnRestart = true,
+                AccountSecurityAction.UnlockWithPinToggle(
+                    UnlockWithPinState.Enabled(
+                        pin = "1234",
+                        shouldRequireMasterPasswordOnRestart = true,
+                    ),
                 ),
             )
         }
